realbackend/requirements.txt
asn1crypto==0.24.0
backports.csv==1.0.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.15
django-ckeditor==5.7.1
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
pipenv==2018.11.26
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
tablib==0.13.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
